+++ rules/firestore/collections.js
@@ -1,0 +1,15 @@
+exports = module.exports = {
+  fl_collections: [
+    'fl_content',
+    'fl_environments',
+    'fl_files',
+    'fl_folders',
+    'fl_locales',
+    'fl_permissions',
+    'fl_schemas',
+    'fl_settings',
+    'fl_users',
+  ],
+  fs_app_collections: ['fs_crop_scores', 'fs_crop_score_cms_link'],
+  fs_service_collections: ['fs_score_info', 'fs_content_errors'],
+};
